Thinking about a simple way to do data-parallel
World is split into "regions" (e.g. boxes), each Chare is a region
with some piece of data in it and they know each other's bounds. Chares
can then send rays to each other as needed. This probably won't be too
fast but it should work and run into some of the harder issues of
how to send rays at all.