Round-robin tile ownership, "distributed" tile rendering
Basic functionality in for rendering distributed tiles, compositing them
based on depth on master after getting all expected tiles for some
region. Each tile is owned by some Chare who is also responsible for
sending the count of regions that project to the tile (the # of regions
main should expect to report a tile).