Project regions to screen to determine who should render tiles
It seems like the box projection hits a wider screen region than I would
expect it too, but probably okish? Unsure why..

Next is to render rays only for pixels where our region is first, and
continuing rays on to the background objects to render the distributed
data. To really test though need to move some of the spheres around so
they actually overlap each other.