Fix one bug with shadow rays
we were incorrectly assuming that if got a hit (a shadow ray was
created) this was the first hit, when it might not be the case. So now
all nodes will generate and send primary rays and we let the depth
compositing sort it out. The other option, which I think is a lot worse,
would be to have all the potential primary hit nodes find what the first
hit is, this would need a lot more communication and generally be a
complete PITA.