Fix bug with primary rays that find a hit but continue on to later boxes
We would incorrectly trample the previous nodes hit result if the later
node didn't find a hit. This is a bit of a mix up in how I'm
storing/sending the depth values that I have to clean up. Right now it's
a bit too easy to lose it or get mixed up in active_ray.ray.t_max
vs. active_ray.color.w when they should always be the same.