Update data paralle to work on splitting intersection from shade, bugs
It seems that the chare is sending its primary ray result message to
itself twice, which makes no fucking sense. IT's the exact same message
ID and everything, so I'm not sure if it's me somehow double-sending it
or Charm++ doing some crazy bullshit related to sending yourself a
message.